:bug: Ensure that image uploads via drop work
- Also reorganized the whole preload bundle, that stuff was getting
messy

Closes #163
